+++ src/ExtendedSeleniumLibrary/decorators/docstring.py
@@ -1,0 +1,51 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+#    Extended Selenium 3 Library - a web testing library with AngularJS support.
+#    Copyright (c) 2015, 2016 Richard Huang <rickypc@users.noreply.github.com>
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU Affero General Public License as
+#    published by the Free Software Foundation, either version 3 of the
+#    License, or (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU Affero General Public License for more details.
+#
+#    You should have received a copy of the GNU Affero General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+Extended Selenium 3 Library - a web testing library with AngularJS support.
+"""
+
+import types
+
+def inherit_docs(cls):
+    """
+    This will copy all the missing documentation for methods from the parent classes.
+
+    :param type cls: class to fix up.
+    :return type: the fixed class.
+    """
+    for name, func in vars(cls).items():
+        if isinstance(func, types.FunctionType) and not func.__doc__:
+            for parent in cls.__bases__:
+                parfunc = getattr(parent, name, None)
+                if parfunc and getattr(parfunc, '__doc__', None):
+                    func.__doc__ = parfunc.__doc__
+                    break
+        elif isinstance(func, property) and not func.fget.__doc__:
+            for parent in cls.__bases__:
+                parprop = getattr(parent, name, None)
+                if parprop and getattr(parprop.fget, '__doc__', None):
+                    newprop = property(fget=func.fget,
+                                       fset=func.fset,
+                                       fdel=func.fdel,
+                                       doc=parprop.fget.__doc__)
+                    setattr(cls, name, newprop)
+                    break
+
+    return cls