+++ src/ExtendedSeleniumLibrary/keywords/extendedselectelement.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-#    Extended Selenium2 Library - a web testing library with AngularJS support.
+#    Extended Selenium 3 Library - a web testing library with AngularJS support.
 #    Copyright (c) 2015, 2016 Richard Huang <rickypc@users.noreply.github.com>
 #
 #    This program is free software: you can redistribute it and/or modify
@@ -18,17 +18,17 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 """
-Extended Selenium2 Library - a web testing library with AngularJS support.
+Extended Selenium 3 Library - a web testing library with AngularJS support.
 """
 
-from Selenium2Library.keywords import _SelectElementKeywords
+from SeleniumLibrary.keywords import SelectElementKeywords
 
 
-class ExtendedSelectElementKeywords(_SelectElementKeywords):
+class ExtendedSelectElementKeywords(SelectElementKeywords):
     """ExtendedSelectElementKeywords are select element execution in the requested browser."""
 
-    def __init__(self):
-        super(ExtendedSelectElementKeywords, self).__init__()
+    def __init__(self, ctx):
+        super(ExtendedSelectElementKeywords, self).__init__(ctx)
 
     def select_all_from_list(self, locator):
         """Selects all values from multi-select list identified by ``locator``.