+++ test/utest/test_extendedjavascript.py
@@ -24,9 +24,9 @@ Extended Selenium2 Library - a web testing library with AngularJS support.
 from sys import path
 path.append('src')
 import unittest
-import mock
-from ExtendedSelenium2Library.keywords import ExtendedJavascriptKeywords
-from Selenium2Library.keywords import _JavaScriptKeywords
+from unittest import mock
+from ExtendedSeleniumLibrary.keywords import ExtendedJavascriptKeywords
+from SeleniumLibrary.keywords import JavaScriptKeywords
 
 
 class ExtendedJavascriptTests(unittest.TestCase):
@@ -34,10 +34,11 @@ class ExtendedJavascriptTests(unittest.TestCase):
 
     def setUp(self):
         """Instantiate the extended Javascript class."""
+        ctx = mock.Mock()
         self.js_code = 'return true'
         self.js_code_py = 'return True'
         self.js_code_vars = 'return ${true}'
-        self.script = ExtendedJavascriptKeywords()
+        self.script = ExtendedJavascriptKeywords(ctx)
         # pylint: disable=protected-access
         self.script._current_browser = mock.Mock()
         self.script._debug = mock.Mock()
@@ -45,7 +46,7 @@ class ExtendedJavascriptTests(unittest.TestCase):
 
     def test_should_inherit_keywords(self):
         """Extended Javascript instance should inherit Selenium2 Javascript instances."""
-        self.assertIsInstance(self.script, _JavaScriptKeywords)
+        self.assertIsInstance(self.script, JavaScriptKeywords)
 
     def test_execute_async_js_with_replaced_vars(self):
         """Should execute async js with replaced vars."""