+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+0.9.2 (2017.11.27)
+==================
+
+- Support Selenium Library 3.0.0+
+
 0.9.1 (2016.07.10)
 ==================
 
@@ -136,11 +141,11 @@
 0.4.9 (2015.07.08)
 ==================
 
-* Removes part of ``_input_text_into_text_field`` method and inherits directly from Selenium2Library
+* Removes part of ``_input_text_into_text_field`` method and inherits directly from SeleniumLibrary
 * Overrides ``get_location`` method with cross browser support
 * Removes ``location_should_be`` method
 * Removes ``_is_firefox`` method
-* Inherits class documentation from Selenium2Library with minor adjustments
+* Inherits class documentation from SeleniumLibrary with minor adjustments
 * Adds documentation and licensing information
 * Removes unneeded exclude in manifest file
 * Minor syntax adjustments